sensor-py/requirements.txt
wheel
smbus2
awsiotsdk
boto3